requirements.txt
boto3>=1.34.0
pandas>=2.0.0
pyarrow>=14.0.0
s3fs>=2024.2.0,<2025.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
apache-airflow>=2.8.0
pyyaml>=6.0
cryptography>=42.0.0
